organizing the notifications methods into a new file and adding the update and delete methods
